requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncio-dgram==2.1.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
croniter==6.0.0
discord-py-interactions==5.14.0
discord-typings==0.9.0
distro==1.9.0
dnspython==2.6.1
emoji==2.14.1
exceptiongroup==1.2.0
filelock==3.13.3
fireworks-ai==0.15.1
frozenlist==1.4.1
fsspec==2024.3.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.4
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.7
idna==3.6
interactions-tasks==1.0.0
Jinja2==3.1.3
jiter==0.5.0
MarkupSafe==2.1.5
mcstatus==11.0.0rc3
mechanize==0.4.10
multidict==6.0.5
nodeenv==1.9.1
numpy==1.26.4
openai==1.45.0
packaging==24.0
pillow==10.4.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.2
PyNaCl==1.5.0
PyOgg @ git+https://github.com/TeamPyOgg/PyOgg
pyright==1.1.397
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
riotwatcher==3.2.5
safetensors==0.4.2
six==1.16.0
sniffio==1.3.1
tenacity==8.2.3
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.2.1
tqdm==4.66.2
transformers==4.44.2
typing_extensions==4.12.2
urllib3==2.2.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
yarl==1.9.4
